+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.11.0 (DEV)
 
+- Add: "Straight line" drawing tool.
 - Enh: Replace number inputs and selects in Maze and Random tools with sliders.
 - Enh: Selection Editors now remember theirs values.